Separate widgets and fairdata from nanopub functions
