+++ fairworkflows/fairworkflows/__init__.py
@@ -1,4 +1,7 @@
 from ._version import __version__ 
-from .search import search, nanosearch, nanofetch
+from .fairdata import FairData
+from .nanopub import nanosearch, nanofetch
 from .step import Step
 from .load import Load
+from .fairwidgets import search
+
